server/python/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pandas==2.1.4
numpy==1.26.3
growwapi==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiohttp==3.9.1
asyncpg==0.29.0
httpx==0.26.0
